Fix build with rust alpha

+++ src/common/signals.rs
@@ -9,10 +9,10 @@
 
 #![allow(dead_code)]
 
-pub static DEFAULT_SIGNAL:uint = 15;
+pub static DEFAULT_SIGNAL:usize= 15;
 
 
-pub struct Signal<'a> { pub name:&'a str, pub value: uint}
+pub struct Signal<'a> { pub name:&'a str, pub value: usize}
 
 /*
 
@@ -139,7 +139,7 @@ pub static ALL_SIGNALS:[Signal<'static>; 31] = [
     Signal{ name: "USR2",   value:31 },
 ];
 
-pub fn signal_by_name_or_value(signal_name_or_value: &str) -> Option<uint> {
+pub fn signal_by_name_or_value(signal_name_or_value: &str) -> Option<usize> {
     if signal_name_or_value == "0" {
         return Some(0);
     }
@@ -153,6 +153,6 @@ pub fn signal_by_name_or_value(signal_name_or_value: &str) -> Option<uint> {
 }
 
 #[inline(always)]
-pub fn is_signal(num: uint) -> bool {
+pub fn is_signal(num: usize) -> bool {
     num < ALL_SIGNALS.len()
 }
