+++ src/hostid/hostid.rs
@@ -1,4 +1,5 @@
 #![crate_name = "hostid"]
+#![allow(unstable)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -33,7 +34,7 @@ mod util;
 static NAME:     &'static str = "hostid";
 static VERSION:  &'static str = "0.0.1";
 
-static EXIT_ERR: int = 1;
+static EXIT_ERR: isize = 1;
 
 pub enum Mode {
     HostId,
