+++ src/timeout/timeout.rs
@@ -1,4 +1,5 @@
 #![crate_name = "timeout"]
+#![allow(unstable)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -32,7 +33,7 @@ extern {
 static NAME: &'static str = "timeout";
 static VERSION: &'static str = "1.0.0";
 
-static ERR_EXIT_STATUS: int = 125;
+static ERR_EXIT_STATUS: isize = 125;
 
 pub fn uumain(args: Vec<String>) -> isize {
     let program = args[0].clone();
@@ -100,7 +101,7 @@ Usage:
     0
 }
 
-fn timeout(cmdname: &str, args: &[String], duration: f64, signal: uint, kill_after: f64, foreground: bool, preserve_status: bool) -> int {
+fn timeout(cmdname: &str, args: &[String], duration: f64, signal: usize, kill_after: f64, foreground: bool, preserve_status: bool) -> isize {
     if !foreground {
         unsafe { setpgid(0, 0) };
     }
@@ -128,7 +129,7 @@ fn timeout(cmdname: &str, args: &[String], duration: f64, signal: uint, kill_aft
             ExitSignal(stat) => stat
         },
         Err(_) => {
-            return_if_err!(ERR_EXIT_STATUS, process.signal(signal as int));
+            return_if_err!(ERR_EXIT_STATUS, process.signal(signal as isize));
             process.set_timeout(Some((kill_after * 1000f64) as u64));
             match process.wait() {
                 Ok(status) => {
@@ -146,7 +147,7 @@ fn timeout(cmdname: &str, args: &[String], duration: f64, signal: uint, kill_aft
                         // XXX: this may not be right
                         return 124;
                     }
-                    return_if_err!(ERR_EXIT_STATUS, process.signal(signals::signal_by_name_or_value("KILL").unwrap() as int));
+                    return_if_err!(ERR_EXIT_STATUS, process.signal(signals::signal_by_name_or_value("KILL").unwrap() as isize));
                     process.set_timeout(None);
                     return_if_err!(ERR_EXIT_STATUS, process.wait());
                     137
